'202406080814am'

+++ src/Sepehr.Infrastructure.Persistence/Repositories/OrderRepositoryAsync.cs
@@ -349,17 +349,22 @@ namespace Sepehr.Infrastructure.Persistence.Repositories
             var ord = _dbContext.Orders
                 .Include(i=>i.Details).ThenInclude(i=>i.PurchaseOrder)
                 .FirstOrDefault(o => o.Id == order.Id);
+            if (ord == null)
+                throw new ApiException("سفارش یافت نشد !");
 
-            var toRemoveDetails = _dbContext.OrderDetails.Where(s => s.OrderId == order.Id && s.WarehouseId != 3 && !order.Details.Select(d => d.Id).Contains(s.Id);
+            var toRemoveDetails = _dbContext.OrderDetails.Where(s => s.OrderId == order.Id && s.WarehouseId != 3 && !order.Details.Select(d => d.Id).Contains(s.Id));
 
             _dbContext.OrderDetails.RemoveRange(toRemoveDetails);
             _dbContext.OrderServices.RemoveRange(_dbContext.OrderServices.Where(s => s.OrderId == order.Id && !order.OrderServices.Select(d => d.Id).Contains(s.Id)));
             _dbContext.OrderPayments.RemoveRange(_dbContext.OrderPayments.Where(s => s.OrderId == order.Id && !order.OrderPayments.Select(d => d.Id).Contains(s.Id)));
 
-            foreach (var oitem in order.Details.Where(d => d.WarehouseId == 3))//.GroupBy(g=> new { g.ProductBrandId,g.Id}))
+            foreach (var oitem in ord.Details.Where(d => d.WarehouseId == 3))//.GroupBy(g=> new { g.ProductBrandId,g.Id}))
             {
                 _dbContext.PurchaseOrder.Remove(_dbContext.PurchaseOrder.First(p => p.Id == oitem.PurchaseOrderId));
-                
+                var inv = await _dbContext.ProductInventories
+                    .FirstOrDefaultAsync(x => x.ProductBrandId == oitem.ProductBrandId && x.WarehouseId == 3);
+
+                inv.ApproximateInventory -= oitem.ProximateAmount;
             }
 
             #region بررسی می شود که کالای مورد ویرایش اگر دارای بارگیری باشد، مقدار بارگیری شده از مقدار اصلی کمتر نباشد
